Note on weak force
Wanted to clear this with you before I did it, but for the weak force we would have to put it within the get_fields2 function because it's dependent on dist. However, this then impacts the field output. I would probably then treat the weak force as a "field" and then later to get acceleration do q/m times that to be consistent with everything else. Let me know what you think! (ignore current location of weak force accel for the time being).